soft delete and rekated changes

+++ src/main/java/rafikibora/services/UserService.java
@@ -15,12 +15,12 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import rafikibora.dto.*;
+import rafikibora.exceptions.ResourceNotFoundException;
 import rafikibora.model.users.User;
 import rafikibora.repository.RoleRepository;
 import rafikibora.repository.UserRepository;
 import rafikibora.security.util.exceptions.RafikiBoraException;
 
-import java.util.List;
 import java.util.Set;
 
 @Service
@@ -103,10 +103,21 @@ public class UserService implements UserServiceI {
         return user.toUserSummary();
     }
 
-    @Override
-    public String deleteUser(int id) {
-        userRepository.deleteById((long) id);
-        return "user removed !! " + id;
+//    @Override
+//    public String deleteUser(int id) {
+//        userRepository.deleteById((long) id);
+//        return "user removed !! " + id;
+//    }
+
+    @Transactional
+    public ResponseEntity deleteUser(int id) {
+        if (userRepository.findById((long) id).isPresent()) {
+            userRepository.deleteById((long) id);
+            // return "account disabled";
+            return new ResponseEntity("User Deleted", HttpStatus.OK);
+        } else {
+            throw new ResourceNotFoundException("User " + id + " Not Found");
+        }
     }
 
     @Override
